Moved Home components into Pages directory
